+++ src/app/components/header/header.component.sass
@@ -1,0 +1,11 @@
+
+.toolbar-left
+  margin-left: auto
+
+
+
+
+
+
+
+
