+++ src/app/components/jokes/jokes.component.html
@@ -1,0 +1,8 @@
+<div class="container">
+  <mat-card class="joke-card">
+    <h1 [className]="show ? 'fade-in joketext' : 'hidden'">{{joke}}</h1>
+  </mat-card>
+  <mat-card class="joke-card-2">
+    <button (click)="getJoke($event)" color="primary" mat-raised-button>Give me a Joke!</button>
+  </mat-card>
+</div>
