+++ src/app/components/jokes/jokes.component.sass
@@ -1,0 +1,47 @@
+.container
+  height: 600px
+  display: flex
+  flex-direction: column
+  justify-content: center
+  align-items: center
+
+
+.joke-card
+  width: 70%
+  display: flex
+  flex-direction: column
+  justify-content: center
+  align-items: center
+  height: 60%
+  opacity: 1
+
+  box-shadow: (none !important)
+
+.joke-card-2
+  display: flex
+  flex-direction: column
+  justify-content: center
+  align-items: center
+  height: 10%
+
+  box-shadow: (none !important)
+
+.joketext
+  text-align: (center !important)
+
+.fade-in
+  opacity: 1
+  animation-name: fadeInOpacity
+  animation-iteration-count: 1
+  animation-timing-function: ease-in
+  animation-duration: 0.25s
+
+.hidden
+  opacity: 0
+
+@keyframes fadeInOpacity
+  0%
+    opacity: 0
+
+  100%
+    opacity: 1
