--- src/app/header/header.component.html
@@ -1,9 +1,0 @@
-<mat-toolbar color="primary">
-	<mat-toolbar-row>
-		<span>Chuck Jokes</span>
-		<div class="toolbar-left">	
-			<a mat-button href="jokes" matTooltip="Every Chuck Norris joke is a five star joke just because it says Chuck Norris"> Jokes </a>
-			<button mat-button  matTooltip="Dev with love by Sebastian Acuña sebastianacuna.it@protonmail.com"> About </button>
-		</div>
-	</mat-toolbar-row>
-</mat-toolbar>
