--- src/app/header/header.component.sass
@@ -1,15 +1,0 @@
-
-
-.toolbar-left
-	margin-left: auto
-
-.header-toolbar-icon-heart
-	margin-top:10px
-
-
-
-
-
-
-
-
