--- src/app/jokes/jokes.component.html
@@ -1,8 +1,0 @@
-<div class="container">
-	<mat-card class="joke-card">
-		<h1 [className]="show ? 'fade-in joketext' : 'hidden'" >{{joke}}</h1>
-	</mat-card>
-	<mat-card class="joke-card-2">
-		<button (click)="getJoke($event)" mat-raised-button color="primary">Give me a Joke!</button>
-	</mat-card>
-</div>
